Done with contact and about
Finished both the about and the contact page. 

Was some debugging to do with a extra white line at the contact page. It was because of a margin bottom, but i didn't really understand it. 

Also learned to use "*{border:1px solid red;}" on the stylesheet to discover all css elements on the page.